+++ calculation/methods/simple_iterations.py
