+++ main.py
@@ -3,16 +3,13 @@ from datetime import datetime as dt
 from numpy import set_printoptions
 
 from visibility.visibility import visibility
-from create_folder import create_folder
 
 
 def main():
     time_start = dt.now()
     print(f"[+] Output image resolution: {n}x{n} pixels.")
 
-    folder = create_folder((n, n), cycle_number)
-
-    visibility((n, n), cycle_number, folder)
+    visibility((n, n), cycle_number)
 
     print(f"[+] Exiting program.\n"
           f"[+] Execution took: {(dt.now() - time_start).total_seconds()}")
